chore: add PEP 765 ast-grep linting rules and fixes (#14816)
## Description

[PEP 765](https://peps.python.org/pep-0765/) introduced a SyntaxWarning
for return/break/continue statements in a finally clause.

This PR adds a new ast-grep linting rule to check for these, as well as
fix the existing occurrences in the codebase.

## Testing

- `scripts/ddtest hatch run lint:sg-test`
  - Validate the rule + the tests for the rule
- `scripts/ddtest hatch run lint:sg`
  - Validate the codebase against our rules

## Risks

There is a small risk I refactored the control flow wrong, please be
sure to review carefully each change. Tests should help validate, but
just in case... worth a quick pause.

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->

---------

Co-authored-by: Emmett Butler <emmett.butler321@gmail.com>