chore(profiling): fix hanging 'test_uwsgi_threads_processes_no_primar… (#14832)
https://datadoghq.atlassian.net/browse/PROF-12706

## Description

`test_uwsgi.py::test_uwsgi_threads_processes_no_primary_lazy_apps` test
is broken in `main` currently (reliably repro-ed on `main`). The symptom
is that it never completes, i.e. appears "hanged" in the terminal.

It seems that only this Python version is affected:
`#1761 10ba61c profile-v2 ... 'runtime.version': '3.13.5'` (from the
logs below)

https://gitlab.ddbuild.io/DataDog/apm-reliability/dd-trace-py/-/jobs/1170947755

It appears that it's impossible to kill the child worker process
(`worker_pid`), which was spawned by the master `uwsgi` (`parent_pid`)
process (even with `SIGKILL`).

After spending some time debugging this, I'm convinced that the
following happens:
1. `parent_pid` is killed and waited on successfully
2. `worker_pid` is cleaned up by the parent, and becomes a zombie during
(1) and so cannot be killed. It is probably in a weird state where it
gets owned by the `init` process and hasn't been reaped yet.
3. `worker_pid` is still in the proc table, so checks with
`kill(child_pid, 0)` succeeds, but ...
4. ... we get trapped in an infinite loop running `kill(child_pid, 0)`
on `worker_pid`

## Testing

Local test:
```
$ ./scripts/ddtest riot -v run --pass-env 10ba61c -- -k test_uwsgi_threads_processes_no_primary_lazy_apps
...
-------------------summary-------------------
✓ profile-v2: [10ba61c] _DD_CIVISIBILITY_USE_CI_CONTEXT_PROVIDER=1 DD_TESTING_RAISE=1 DD_REMOTE_CONFIGURATION_ENABLED=false DD_INJECTION_ENABLED=1 DD_INJECT_FORCE=1 DD_PATCH_MODULES=unittest:false CMAKE_BUILD_PARALLEL_LEVEL=12 CARGO_BUILD_JOBS=12 DD_PROFILING_ENABLE_ASSERTS=1 CPUCOUNT=12 pythonInterpreter(_hint='3.13') 'mock' 'pytest' 'pytest-mock' 'coverage' 'pytest-cov' 'opentracing' 'hypothesis<6.45.1' 'gunicorn' 'jsonschema' 'lz4' 'pytest-cpp' 'pytest-benchmark' 'py-cpuinfo~=8.0.0' 'pytest-asyncio==0.21.1' 'pytest-randomly' 'uwsgi' 'protobuf'
1 passed with 0 warnings, 0 failed
```

## Risks

none

## Additional Notes

n/a