fix(ray): runtime_env can be None leading to a job error (#14837)
When submitting a job, `runtime_env` could be none in `kwargs`, leading
to an `AttributeError`, see this
[link](https://dd.datad0g.com/llm/distributed-ai/jobs?query=%40component%3Aray%20parent_id%3A0%20job-113464b6-33d0-47&agg_m=count&agg_m_source=base&agg_t=count&colorLegendSort=time&fromUser=false&spanId=5240087393744960581&traceId=68e818d800000000e5341ed29f2823d9&start=1759437349664&end=1760042149664&paused=false).

This PR fixes this issue. It also updates two snapshots that were
missing a span. The tests were xfailing so this was not caught in CI.

(cherry picked from commit f11c438efa31ade5b13a5b063863f675ea5c687c)