feat(ci_visibility): use sys.monitoring code coverage for Python 3.13 (#14786)
## Description

Currently code coverage support is a no-op for Python 3.13. It should
support the same mechanism used for 3.12 (sys.monitoring). This PR
enables that.

## Testing

The existing unit tests for coverage should pass for Python 3.13.

## Risks

It is possible that differences in 3.13 bytecode cause lines not to be
detected correctly; the unit tests should catch chat. sys.monitoring
code coverage does not depend on bytecode _modification_ but it does
depend on _parsing_ the bytecode to identify instructions of interest,
so differences in the bytecode instructions used might affect the
results.

## Additional Notes

It turns out that for Python 3.14 some bytecode instructions _are_
different. I will address it in a separate PR.