ci: fix scripts/run-tests issues (#14858)
## Description

This PR fixes a few small issues with the scripts/run-tests helper.

1. No long requires trailing slash when selecting from a directory
2. Service selection
a. Previously would collect all possible services needed from all
possible test suites, not will limit to only the services needed for the
selected venvs, and will start only the necessary services before each
venv run, then tear them down when it is done, before starting the
services needed for the next venv run, and so on.

## Testing

<!-- Describe your testing strategy or note what tests are included -->

## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->