fix(grpc): leaking spans when using gRPC future interface (#14875)
We've tracked down some issues related to broken span hierarchy down to
the way that the gRPC instrumentation deals with entering spans, which
causes the grpc span to leak outside it's call when using the .future()
interface. This PR changes the instrumentation to explicitly only enter
the span when calling the continuations instead of just entering
immediately.

## Description

<!-- Provide an overview of the change and motivation for the change -->

## Testing

I've added tests, but haven't run them for a while and working on
getting it running locally. Will update if I run into any issues

## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

## Additional Notes

---------

Signed-off-by: Sam Schlegel <sam.s@discordapp.com>
Co-authored-by: Sam Schlegel <sam.s@discordapp.com>