chore: clean-up old Span._resource mutable pattern (#14861)
## Description

We used to store resource name as a list so that the profiling product
could grab a mutable reference to the resource name since it may change
during the lifetime of a trace.

This hack is no longer used, we can go back to storing resource name as
just a normal string.

## Testing

<!-- Describe your testing strategy or note what tests are included -->

## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->