fix(ddwaf): allow conversion of arbitrary Mappings and Sequences to ddwaf_object (#14913)
## Description

Allow conversion of arbitrary Mappings and Sequences to ddwaf_object
type.

## Motivation

When using `requests` without `urllib3` patching with `urllib3<2`,
headers are represented by a `requests.structures.CaseInsensitiveDict`
instead of a regular Dict when they are intercepted at the `http.client`
level by the patched `HTTPConnection.request` function. When converting
to ddwaf_object this coerces it into a string instead of properly
representing it as a `map`. This makes headers matching effectively
useless.

## Testing

- Added a unit test

## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->