feat(llmobs): add reasoning for custom evals (#14919)
## Description
(public change) Adds `reasoning` as an argument to
`submit_evaluation_for()` and `submit_evaluation()`. This arg is used to
denote an explanation behind the evaluation results (i.e. why was the
span marked as toxic?)

(internal change - not facing users) Also changes how the `assessment`
field is stored on the evaluation object (#14792 added it as a nested
`success_criteria` object) to a top-level field on the evaluation
object. This isn't breaking (since this hasn't been officially released
on our product backend) nor a user-facing change.

<!-- Provide an overview of the change and motivation for the change -->

## Testing

<!-- Describe your testing strategy or note what tests are included -->

## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->