chore: refactor tracer partial flushing logic (#14860)
## Description

This simplify the code and removes the edge case where we tag a complete
trace chunk with unnecessary partial flush tag

We move some of the complexity of removing finished spans from a trace
into a method.

We add __slots__ to _Trace object to aid in memory usage and size.

Small performance win to use `span.duration_is is not None` to check if
a span if finished instead of using the `finished` property function.

## Testing

<!-- Describe your testing strategy or note what tests are included -->

## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->