chore(core): deprecate core.dispatch_with_results (#14915)
## Description

We want to move people away from `core.dispatch_with_results`.

This PR adds a linting rule to check for usage and encourage other
patterns with `core.dispatch`.

We added explicit allows for the existing usage and we can work to
slowly migrate them away.

## Testing

Added tests for the ast-grep linting rules.

## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->