ci: increase otel-span RSS SLOs (#14976)
## Description

Memory usage on `main` is just above the threshold. This PR increases
the SLO for now.

## Testing

<!-- Describe your testing strategy or note what tests are included -->

## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

## Additional Notes

The Span add metric/tag benchmarks also use a ton of memory, their SLOs
are almost 1gb. We are probably not clearing spans from span aggregator
or something between loops, meaning the amount of memory used is
dependent on how many loops/iterations pyperf decides to run and not the
actual memory usage of the library.