fix(llmobs): ensure langchain azure openai spans are not duplicate llm marked (#14939)
[MLOB-4230]
## Description
This PR does 3 things:
1. (non-user facing) Updates our docker-compose and services.yml files
to upgrade to the latest testagent version, as well as adding a env var
`VCR_PROVIDER_MAP` value for the testagent configs.
2. (user-facing) fixes the langchain integration such that azure openai
calls are not marked as duplicate LLM spans (if the openai integration
is enabled), and instead marks them as generic workflow spans.
3. (non-user facing) Adds langchain tests for calling Azure OpenAI.
These requires the testagent upgrade and the `VCR_PROVIDER_MAP` env var
to allow the testagent vcr proxy to call the azure openai endpoint.

We have logic in our langchain integration to mark specific LLM calls as
generic workflow spans (instead of the default llm span) if we detect
the corresponding integration (for the given provider, i.e.
`openai/anthropic`) is also enabled and will result in a downstream LLM
span. Our product experience breaks if multiple spans duplicate
represent an LLM call, and we were previously missing support for azure
openai.

<!-- Provide an overview of the change and motivation for the change -->

## Testing

<!-- Describe your testing strategy or note what tests are included -->

## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->


[MLOB-4230]:
https://datadoghq.atlassian.net/browse/MLOB-4230?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ