fix(tracer): ensure compatibility with wrapt 2.0 (#14959)
## Description

wrapt 2.0 released in 2025 oct 20 has some breaking changes.

In particular, BaseObjectProxy is now the base class of all object
proxy.

Also: typing was introduced in wrapt 2.0, but this is breaking our mypy
steps, because it was done too naively. This PR makes the necessary
changes to pass this step.

## Testing

This is tested with all existing tests and should fix new failures on
system tests related to this change.


APPSEC-59731