chore: improve git worktree support with scripts/ddtest (#15028)
## Description

`scripts/ddtest` + git worktrees don't play nicely together. This is
because you get the local/cwd mounted into the container which means you
will likely be missing the real `.git` directory (as well, worktrees
love setting absolute paths in the git config).

This means things like `pip install -e .` won't work because
setuptools_scm cannot query git to figure out the version.

With this change when running `scripts/ddtest` we will try to determine
if we are in a worktree, and if yes, then compute the setuptools_scm
version outside of the container and then inject it into it.

If the script/automation doesn't work for some, we will passthrough
`SETUPTOOLS_SCM_PRETEND_VERSION` and
`SETUPTOOLS_SCM_PRETEND_VERSION_FOR_DDTRACE` env variables so the user
could always manually inject the version they care about.

## Testing

<!-- Describe your testing strategy or note what tests are included -->

## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->