perf(profiling): do slightly less work when possible (#15025)
## Description

This PR updates a small part of the logic in the Profiler / Stack V2 to
only fetch the `filename_str` if we are going to use it (which may not
be the case if we are pushing a "Task name frame".

## Risks

As far as I can tell, this is a functional no-op as it is effectively
only deferring the declaration (and populating) of a variable –
`filename_str` – to later in the code.