fix(internal): fix for potential race condition (#15018)
## Description

We had a potential (and reported to happen to a client) race condition
where the `self.endpoints` list in `internal/endpoints.py` could change
size during iteration (modified by a thread), causing an exception.

This fixes it by using a tuple copy of the iterated list and adds
regression tests.

Thanks Adria Ardilla for reporting and suggesting the fix.

---------

Signed-off-by: Juanjo Alvarez <juanjo.alvarezmartinez@datadoghq.com>