feat: remove Python 3.8 support, deprecate Python 3.9 support (#14891)
This change removes support and testing for Python 3.8, which reached
its end-of-life in late 2024 and was deprecated in ddtrace in
https://github.com/DataDog/dd-trace-py/pull/14890. It also deprecates
support for Python 3.9, which reached its end-of-life recently.

This includes removal of the `cassandra` integration, which is only
tested under Python 3.8.

Note that the base branch is not `main`, but a branch on which we can
stage the breaking changes for the 4.0 release.

Depends on https://github.com/DataDog/system-tests/pull/5547

---------

Co-authored-by: Sam Brenner <106700075+sabrenner@users.noreply.github.com>
Co-authored-by: Gabriele N. Tornetta <P403n1x87@users.noreply.github.com>