chore(ci_visibility): retry Test Optimization API calls on server errors (#14950)
## Description

Currently, if a request to the Test Optimization settings API fails with
a 500 or other server error, we do not try again, and run the test
session with Test Optimization features disabled. This PR adds retries
to the Test Optimization API requests, for server errors (status >= 500;
it does not make sense to retry 4xx errors as they are unlikely to
succeed if called again with the same parameters) and network errors
(such as timeout).

## Testing

Unit tests added.

## Risks

None.

## Additional Notes

None.