feat(api-security): use simplified endpoint for API Security Sampling when route is missing (#14960)
## Description

In some instrumented setups, the REQUEST_ROUTE waf address might not be
available. This is the case with:
- lambda function behind Application Load Balancer
- lambda function behind lambda function url services

To still compute and tag API Schemas in this case, we can sample using
the simplified endpoint computed by trace resource renaming or compute
it if trace resource renaming is disabled.

To avoid unnecessarily computing the endpoint on spans that correctly
miss the route, we keep the route empty when:
- the request is a 404
- the request has been blocked

## Testing

- Added unit tests in API Security Manager

## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->