chore(ci_visibility): file-level coverage for py3.12+ (#15081)
## Description

CI Visibility: Add a faster mode for code coverage intended for TIA,
that only would track file paths for the executed python code and not
per-line coverage inside them.

It works only for Python 3.12+, where we use the `sys.monitoring` API.
But instead of using `LINE` events, that are triggered for each line of
code executed, if the env var `_DD_COVERAGE_FILE_LEVEL` is set, we would
use `PY_START` events, that are triggered when the code objects start
executing (but not for all their lines)

Disabled by default, but can be enabled with the env var
`_DD_COVERAGE_FILE_LEVEL=true`.

## Testing

Tests are adapted so they are run with both modes, for file level mode
we don't assert the actual lines inside the tests.

## Risks

None

---------

Co-authored-by: Vítor De Araújo <vitor.dearaujo@datadoghq.com>