fix(ci_visibility): do not fetch tags while unshallowing git repo (#15133)
## Description

Currently when calling `git fetch` to unshallow a repo, we end up
fetching repo tags, even though we don't need them. This can have a
performance impact in repos with a large number of tags. This PR changes
it to call `git fetch` with `--no-tags`.

## Testing

Unit tests and manual testing.

## Risks

None.

## Additional Notes

None.