chore(tracing): move meta/metrics type checking to the encoder (#14982)
## Description

The benefit here is we can do a first pass on meta/metrics to filter out
bad values before we try to pack them, this way we can throw out
individual bad values instead of throwing out the whole span/trace.
## Testing

<!-- Describe your testing strategy or note what tests are included -->

## Risks

This will have a negative performance impact for now until #14943 can be
merged. The impact on the flask/django benchmarks is pretty low, so
while this is showing a +20% increase locally for the encoder
benchmarks, the real world impact on the flask/django microbenchmarks is
<1% so even without #14943 it might be "ok" to introduce.


Right now we do type checking in `set_tag`/`set_metric`. With this PR we
will do the type checking there **and** in the encoder. Once #14943
merges we'll remove the type checking from `set_tag`/`set_metric` and
only have it in the encoder. It should have the benefit of using the
C-API for the type checking (faster), and we don't do type checking on
spans that aren't going to get encoded.


Note: the end result could be a tad slower on average still because we
are keeping `for k, v in meta.items():` but then we are adding a `for k,
v in filtered_meta_list:` so the additional iteration could bite us. It
could be worth trying to find a way to pack the meta/metrics without
first knowing the end size.

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->

---------

Co-authored-by: Emmett Butler <emmett.butler321@gmail.com>