fix: modify error tracking exceptions (#15148)
## Description

<!-- Provide an overview of the change and motivation for the change -->

For [APMS-17683](https://datadoghq.atlassian.net/browse/APMS-17683)
customer was having an issue where their custom exception object was
unhashable, resulting in a `TypeError` when trying to use the object as
a key. This fix changes so that we store the exception object id, not
the exception object.

## Testing

<!-- Describe your testing strategy or note what tests are included -->

## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->

[APMS-17683]:
https://datadoghq.atlassian.net/browse/APMS-17683?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

(cherry picked from commit 84db0e6062e72ae4ce3968f25b89628053841d06)