chore(tracing): simplify Span tag/metric API typing (#14943)
## Description

**4.0**

This updates all tag and metric methods to explicitly be `dict[str,
str]` and `dict[str, int float]` only.

This removes the `value: Any` from `set_tag` which allows us to remove a
lot of checks for things like if it is a numeric type or not.

## Testing

<!-- Describe your testing strategy or note what tests are included -->

## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

## Additional Notes

We needed to bump the `otelspan-add-tags` benchmark SLO because we also
needed to add additional processing in the otel span `set_attribute` to
delegate types between `set_tag` and `set_metric`.

---------

Co-authored-by: Emmett Butler <emmett.butler321@gmail.com>
Co-authored-by: Emmett Butler <723615+emmettbutler@users.noreply.github.com>