ci: add `[pyX.Y]` suffix to test names (#15191)
## Description

Workaround for our Flaky Test Management product, since a test that is
flaky for a single Python version would be quarantined for all Python
versions at the moment (as it is quarantined by name).

The workaround consists in adding the python version as a suffix, so the
same test for other Python versions would not be quarantined.

## Testing

Manually tested that all the tests are sent with the suffix.

## Risks

It effectively changes the test names, so it will impact reporting, etc.

It also complicates some bits, as we want the original test name for
snapshots and similar tests.

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->