ci(benchmarks): fix high memory usage in span + otelspan (#15200)
## Description

These benchmarks were using like 900+ mb of RSS.

In the case when we don't finish the spans, the way we were clearing the
spans was in an un-optimal place, since we would build up all created
spans across all loops in memory and then at the end clear them. This
clears them after every loop, so each loops/max memory usage is only the
max memory for that specific loop.

## Testing

<!-- Describe your testing strategy or note what tests are included -->

## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->