feat(profiling): add sample count to internal payload (#15177)
## Description

This PR adds data to the `internal` payload we send to `libdatadog`. In
this `internal` payload, we can push custom metrics (that are not
exposed to customers) but that we can access for analytics.

For the time being, I only added the number of Samples taken (one per
thread) and Sampling Events (one for each time we run `for_each_interp`)
for the current Profile. We may want to add more in the short term (e.g.
number of times adaptive sampling was adjusted, history of adaptive
sampling intervals, etc.)

In order to implement this feature in a way that keeps clear semantics
around locking the Profile object, I refactored the code not to `borrow`
only the single `Profile` but instead both the `Profile` and the
`ProfilerStats` (in which we store our metrics). To keep locking clear
and explicit, locking returns a `ProfileBorrow` objects which allows to
access both the `Profile` and the `ProfilerStats` and that is
RAII-compatible (automatically unlocks when it goes out of scope).

Note the PR does add some "lock contention" because we now need to take
the Profile lock in two new places (one for each metric we increment) –
it should be the same order of magnitude as what we already do though
(in number of lock acquisitions).
We plan to refactor Stack V2 in the short to avoid having to hold the
Profile / ProfilerStats lock during the upload, by using double
buffering or by releasing earlier (after the Profile data has been
serialised). This should improve the performance and reduce lost
Samples.

The PR also adds an integration test, ensuring that the generated JSON
string is correct (we can't guarantee exact numbers, but we can check
the numbers we have are meaningful).

**Note** I marked the PR as `no-changelog` as this feature isn't
public/visible by customers.

Open questions:
*  What other metrics do we want to add?

This is an example of querying through the Events UI.

<img width="886" height="604" alt="image"
src="https://github.com/user-attachments/assets/53c3eae0-d29a-4446-9f1b-72ac29a75d60"
/>