chore: lazy context wrapping (#15295)
## Description

We implement a lazy context wrapping mechanism whereby the bytecode
context wrapping is performed on first invocation of the function. Note
that this still requires the original function to be instrumented via
bytecode manipulations.

## Additional Notes

The motivation for this change is to allow for a more lightweight
bytecode instrumentation to trigger a more expensive one only when
needed. This is typically when an instrumented function is called for
the first time.