ci: reduce flakiness from httplib https tests (#15318)
## Description

We are trying to hit a real third party service, which is now down
because of cloudflare outage.

We don't actually care that the request was successful, at least from a
"200" vs "404" vs "500", we just want to be sure that using the HTTPS
request mechanism does properly create a span.

## Testing

<!-- Describe your testing strategy or note what tests are included -->

## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->