chore(utils): stop threads in dummy writer (#15307)
## Description

Fix crash in `appsec_threats` test on macos caused by to many
NativeWriter being created. The crash is due to each native writer
instance holding onto a kqueue fd exceeding the macos ulimit (256). This
fixes the issue by shutting down the tokio runtime of the native writer
when we are done with the DummyTracer ie. when pop is called.

## Testing

Run the `appsec_threats_fastapi` test with `ulimit -n` set to 256 (by
default on mac).

---------

Co-authored-by: Christophe Papazian <christophe.papazian@datadoghq.com>
Co-authored-by: Christophe Papazian <114495376+christophe-papazian@users.noreply.github.com>