chore(internal): add process_tags to APM payload (#15146)
This PR implements this
[RFC](https://docs.google.com/document/d/1AFdLUuVk70i0bJd5335-RxqsvwAV9ovAqcO2z5mEMbA/edit?pli=1&tab=t.0#heading=h.s9l1lctqlg11)
for the tracing product.

Add process_tags in APM payload. According to the RFC and as the python
tracer supports only v0.4 and v0.5 the tag `_dd.tags.process` is set in
the first first span of each trace chunk set.

The set tag are:
- `entrypoint.workdir`. The working dir name, should be the last path
segment.
- `entrypoint.name`. The entrypoint of the application. It is the the
script name.
- `entrypoint.type`. I'd assume that it can be only script for Python.
- `entrypoint.basedir`. The base dir is the directory where the called
executable resides. Should be the last path segment.