fix(llmobs): do no annotate `self` for automatic annotations with decorators (#15349)
## Description

Makes sure `self` is not decorated when using LLMObs decorators.

Closes #14849 

## Testing

Unit tests added

## Risks

Should not be a breaking change as although we run evaluations on input
values, they are not used as far as I know as part of monitors or
sampling rates for evaluations. If anything, it might help with
evaluations on spans that had `self` annotated previously as a result of
using decorators without manual annotation override.