feat(profiling): make Profile upload lock-free (#15302)
## Description

https://datadoghq.atlassian.net/browse/PROF-13044

This PR reduces locking around the libdatadog Profile object to make
uploading non-blocking for sampling. Previously, the Scheduler thread
(which uploads every once in a while) would take the Lock on the Profile
object, preventing any new Samples from being pushed to it until the
upload finishes. Although we don't have any numbers on that, it is
possible that this would trigger adaptive sampling or make us lose some
data.

The new implementation doesn't make uploading completely lock-free, but
already improves it by eliminating the need for locking _during the
upload_. We still need to lock the `Profile` object when we encode it.
It may be possible to optimise this further by having two `Profile`
objects (one we would be writing to, and one that would be _being
uploaded_) but this requires more complicated lock handling to actually
be safe. (Reason is, having two Profiles is still a problem if the
previous encoding/uploading hasn't finished when the next needs to be
encoded/uploaded.)

In the new implementation, the Profile Lock is acquired just before
encoding is done, and released right after. Profiler Stats are also
copied and reset while the Lock is held.

In order to make that work, I had to slightly adapt the interface of
`Uploader` for it to work an `EncodedProfile` instead of a `Profile`,
which allowed me to hoist the encoding outside of the `Uploader` (to the
`UploaderBuilder`), and thus only hold the Profile Lock during
`UploaderBuilder::Build` (as opposed to during the whole of
`Uploader::upload`).
`ProfilerStats` are captured and reset by `std::swap`-ping them with an
empty `ProfilerStats` object, eliminating the previous "singleton-like"
`ProfilerStats` we had. (This works and is trivial because
`ProfilerStats` is a C++ object; we cannot do the same for the `Profile`
itself, as it is managed by the Rust library.)
`Uploader`'s interface has also slightly changed – it is now constructed
with the `ProfilerStats` and `EncodedProfile` directly (not taken as
argument for `upload`).