ci: update benchmark slo + reporting dependencies (#15339)
## Description

Only run SLO check if the benchmark jobs all succeeded.

Only run the reporting if the SLO check job ran.


Right now if a benchmarks fails (programming error, or infra issues,
etc), we still run the SLO check, which we know will fail because we
don't report all benchmark results, then we also run the benchmark
reporter which will succeed. If you then go back and re-run the failed
benchmark job (assuming flaky infra, for example), then you have to know
to manually re-run the SLO check and the reporting job as well).

This should simplify and clarify some of it a little bit.

## Testing

<!-- Describe your testing strategy or note what tests are included -->

## Risks

<!-- Note any risks associated with this change, or "None" if no risks
-->

## Additional Notes

<!-- Any other information that would be helpful for reviewers -->