chore(profiling): move echion to dd-trace-py (#15136)
https://datadoghq.atlassian.net/browse/PROF-12683

This PR moves the source code of Echion from the Echion repository
(https://github.com/P403n1x87/echion) to dd-trace-py. We hope that doing
this will lead to an improved developer experience in dd-trace-py, as we
will effectively:

- Reduce the number of touch points required when we make a change to
the Python sampler
- Eliminate the need for "round-trips" between Echion and dd-trace-py
(when the dd-trace-py CI captures an issue that the Echion CI missed)
- Make it easier to iterate locally in dd-trace-py (currently,
re-building Stack V2 with an updated version of Echion requires nuking
the `build/` directory / cache, meaning we need to build everything from
scratch, which takes time)

**Note** The version of the code that is in this PR is **not** the
latest version of Echion (missing a few PRs). I am still sending the PR
for review (so that I can get an approval on the form and general idea)
and I will make sure to include/backport any changes merged to Echion
before I merge this PR to dd-trace-py.

**Note** I have copied `.clang-format` over as to minimise the size of
the diff and make it easier to port changes from Echion while the PR is
in review. I will remove the custom `.clang-format` and use the default
one from dd-trace-py in a later PR.

(cherry picked from commit 38cc2e8f24241f0485aa644a400bd3689ef9ebda)